refactor: move methods in stat_tracker to new module season_statable
